+++ frontend/src/components/GameComponents/BiteHuman.js
