+++ frontend/src/components/GameComponents/BiteZombie.js
