+++ frontend/src/components/GameComponents/ChooseFaction.js
