+++ frontend/src/components/SquadComponents/SquadDetail.js
