+++ frontend/src/components/SquadComponents/SquadList.js
