+++ frontend/src/components/StylingComponent/Header.js
