+++ frontend/src/components/StylingComponent/NavBar.js
