+++ frontend/src/components/StylingComponent/NotFound.js
